+++ src/ra/menus.cc
@@ -80,8 +80,8 @@ static int Coordinates_In_Region(int x, int y, int inx1, int iny1, int inx2,
                                  int iny2);
 static int Select_To_Entry(int selection, unsigned long enabled_mask,
                            int start_bit);
-static void Flash_Line(const char* text, int xpix, int ypix, unsigned nfgc,
-                       unsigned hfgc, unsigned bgc);
+static void Flash_Line(const char* text, int xpix, int ypix, int nfgc,
+                       int hfgc, int bgc);
 
 int UnknownKey;
 
@@ -140,8 +140,8 @@ static int Select_To_Entry(int selection, const unsigned long enabled_mask,
 /*	RETURNS:	none
  */
 /*=========================================================================*/
-static void Flash_Line(const char* text, int xpix, int ypix, unsigned nfgc,
-                       unsigned hfgc, unsigned bgc) {
+static void Flash_Line(const char* text, int xpix, int ypix, int nfgc,
+                       int hfgc, int bgc) {
   int loop;
 
   for (loop = 0; loop < 3; loop++) {
@@ -446,7 +446,7 @@ int Do_Menu(const char** strings, bool /*unused*/) {
   length = 0;
   ptr = strings;
   while (*ptr) {
-    length = std::max(length, static_cast<int>(String_Pixel_Width(*ptr)));
+    length = std::max(length, String_Pixel_Width(*ptr));
     ptr++;
   }
   length += 7;
