+++ src/ra/woledit.cc
@@ -48,9 +48,9 @@ void WOLEditClass::Draw_Text(const char* text) {
   Conquer_Clip_Text_Print(text, X + 1, Y + 1, Color, TBLACK, TextFlags | flags,
                           Width - 2);
 
-  const int text_width = static_cast<int>(String_Pixel_Width(text));
+  const int text_width = String_Pixel_Width(text);
   if (Has_Focus() &&  //	strlen(text) < MaxLength &&
-      text_width + static_cast<int>(String_Pixel_Width("_")) < Width - 2) {
+      text_width + String_Pixel_Width("_") < Width - 2) {
     Conquer_Clip_Text_Print("_", X + 1 + text_width, Y + 1, Color, TBLACK,
                             TextFlags | flags);
   }
